+++ Frontend/src/app/_components/admin/Auth/withAuth.tsx
@@ -7,6 +7,7 @@ import { red } from "@mui/material/colors";
 import DataServiceContext from "@/app/_core/services/dataService";
 import { Loading } from "@/app/_core/components/loading/loading";
 import { IHTTPdetail, IHTTPresponse } from "@/app/_core/interfaces/core.interface";
+import { HTTPMethodEnum } from "@/app/_core/enums/core.enum";
 
  //HOC AUTH
 export default function withAuth(Component: any){
@@ -31,7 +32,7 @@ export default function withAuth(Component: any){
 
         const authenticate = async () => {
             const options = api_service.generateOptions(url,{})
-            const response: IHTTPresponse<IHTTPdetail> = await api_service.request<IHTTPdetail>("GET", options, true)
+            const response: IHTTPresponse<IHTTPdetail> = await api_service.request<IHTTPdetail>(HTTPMethodEnum.GET, options, true)
     
             if (response.status != 200){
                 router.push("/admin/login")