+++ Frontend/src/app/_components/admin/chat/interfaces/chat.interface.ts
@@ -1,17 +1,9 @@
-import { IMessageData } from "@/app/_components/chat/interfaces/chat.interface"
+import { IChat, IMessageData } from "@/app/_components/chat/interfaces/chat.interface"
 
 export interface IChatPanelProp{
     chatHandler: (chat:IChat) => void
 }
 
-export interface IChat{
-    created:string
-    roomID: string
-    status: string
-    client: IClient
-    messages: IMessageData
-    
-}
 
 export interface IClient{
     name: string