+++ Frontend/src/app/_components/admin/login/login.tsx
@@ -3,11 +3,12 @@ import DataServiceContext from "@/app/_core/services/dataService"
 import httpRequestContext from "@/app/_core/services/httpRequest"
 import { FormEvent, useContext, useEffect } from "react"
 import loginServiceContext from "./services/login.service"
-import { IHttpOptions } from "@/app/_core/interfaces/core.interface"
+import { IHTTPresponse, IHttpOptions } from "@/app/_core/interfaces/core.interface"
 import { IRequestUser } from "./interfaces/login.interface"
 
 import { useRouter } from "next/navigation"
 import { Loading } from "@/app/_core/components/loading/loading"
+import { HTTPMethodEnum } from "@/app/_core/enums/core.enum"
 
 export default function Login() {
     const api_service = useContext(httpRequestContext)
@@ -21,16 +22,16 @@ export default function Login() {
     const requestLogin = async (data: any) => {
         const requestData: IRequestUser = login_service.processData(data.username, data.password)
         const options: IHttpOptions = api_service.generateOptions<IRequestUser>(url, requestData)
-        const [status, response] = await api_service.request('POST', options)
+        const response: IHTTPresponse<any> = await api_service.request<any>(HTTPMethodEnum.POST, options)
 
-        if(status != 200){
+        if(response.status != 200){
             //toast warning logic
             return
         }
         
-        if(status === 200){
-            router.push("/admin/dashboard")
-        }
+
+        router.push("/admin/dashboard")
+        
     }