+++ Frontend/src/app/_components/admin/stats/interfaces/stats.interface.ts
@@ -1,5 +1,19 @@
 export interface IStats {
+    chatsCount: IChatsCount | undefined
+    todayVisits: ITodayVisits | undefined
+    totalVisits: ITotalVisits | undefined
+}
+
+export interface ITodayVisits {
     date: string
     visited: number
-    total_visits: number
+}
+
+export interface ITotalVisits {
+    total: number
+}
+
+export interface IChatsCount {
+    chats_on: number 
+    chats_off: number
 }