+++ Frontend/src/app/_components/init/init.tsx
@@ -1,4 +1,5 @@
 "use client"
+import { HTTPMethodEnum } from "@/app/_core/enums/core.enum";
 import DataServiceContext from "@/app/_core/services/dataService";
 import httpRequestContext from "@/app/_core/services/httpRequest";
 import { useContext, useEffect } from "react";
@@ -13,7 +14,7 @@ export default function Init(){
     const generateVisit = async () => {
         const url = data_service.setUrl("register_create")
         const options = api_service.generateOptions(url,{})
-        await api_service.request("POST",options)
+        await api_service.request(HTTPMethodEnum.POST,options)
     }
 
     return <></>