+++ Frontend/src/app/_core/enums/core.enum.ts
@@ -2,3 +2,18 @@ export enum WebsocketTypeEnum{
     Chat = 0,
     Admin = 1
 }
+
+export enum HTTPMethodEnum{
+    GET = "GET",
+    POST = "POST",
+    PUT = "PUT",
+    DELETE = "DELETE"
+}
+
+export enum HTTPCodeEnum{
+    OK = 200,
+    CREATED = 201,
+    NOT_FOUND = 404,
+    BAD_REQUEST = 400,
+    FORBIDDEN = 403
+}